train/object-detection/baseline/frameworks/yolo/pyproject.toml
[project]
name = "framework-yolo"
version = "0.1.0"
description = "Standalone YOLO framework runner for the multi-framework orchestrator"
requires-python = ">=3.10"
dependencies = [
    "ultralytics>=8.3.227",
    "pyyaml>=6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
